Spaceship moving and rotation.

+++ Assets/Spaceship/Movement/ISpaceshipMover.cs
@@ -5,7 +5,7 @@ namespace Asteroids.Spaceship.Movement
 {
     public interface ISpaceshipMover
     {
-        int GetThrottleValue();
+        float GetThrottleValue();
         float GetTurnDirectionValue();
     }
 }
